split announce bridge module
